add restriction for not showing add and delete components

+++ albiFront/src/app/components/sublet-delete/sublet-delete.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Sublet } from '@models/sublet';
 import { SubletService } from '@services/sublet.service';
+import { TokenStorageService } from '@services/token-storage.service';
 
 @Component({
   selector: 'app-sublet-delete',
@@ -10,15 +11,24 @@ import { SubletService } from '@services/sublet.service';
 })
 export class SubletDeleteComponent implements OnInit {
 
+  currentUser: any;
+
   constructor(private route: ActivatedRoute, 
-      private router: Router, 
+      private router: Router,
+      private token: TokenStorageService,
       private subletService: SubletService) { }
 
   ngOnInit(): void {
-  this.route.queryParams.subscribe(params => {
-        let id = params['id'];
-        this.subletService.deleteSublets(id).subscribe();
-      });
+    this.currentUser = this.token.getUser();
+        if(!this.currentUser.id){
+          this.router.navigate(['/login/']);  
+        }
+        else{
+          this.route.queryParams.subscribe(params => {
+                let id = params['id'];
+                this.subletService.deleteSublets(id).subscribe();
+           });
+        }
   }
 
 }
